break up record proto
