+++ src/components/Protobuf/RecordProto/EnumRow/index.js
@@ -1,0 +1,3 @@
+import EnumRow from './EnumRow'
+
+export default EnumRow
