+++ src/components/Protobuf/RecordProto/RecordInterface/index.js
@@ -1,0 +1,3 @@
+import RecordInterface from './RecordInterface'
+
+export default RecordInterface
